deleted:    lib/advent2022/day11.txt 	deleted:    lib/advent2022/day11t.txt 	modified:   src/advent2022/day11to15/data.java 	modified:   src/advent2022/day11to15/day11.java 	new file:   src/advent2022/day11to15/day11b.java

+++ src/advent2022/day11to15/data.java
@@ -1,14 +1,15 @@
 package advent2022.day11to15;
 import java.util.ArrayList;
 
+
 public class data {
-    ArrayList<Integer> items;
+    ArrayList<Long> items;
     char operator;
     int operatorNum;
     int divisor;
     int truth, falth;
 
-    public data (ArrayList<Integer> items, char operator, int operatorNum, int divisor, int truth, int falth){
+    public data (ArrayList<Long> items, char operator, int operatorNum, int divisor, int truth, int falth){
         this.items = items;
         this.operator = operator;
         this.operatorNum = operatorNum;
